pyproject.toml
[project]
name = "opencode-a2a"
version = "0.1.0"
description = "A2A wrapper service for opencode"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk==0.3.17",
    "fastapi>=0.110",
    "httpx>=0.27",
    "pydantic>=2.6",
    "sse-starlette>=2.1",
    "uvicorn>=0.29",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=3.7",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.5",
]

[project.scripts]
opencode-a2a = "opencode_a2a.app:main"

[tool.ruff]
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.ruff.format]
quote-style = "double"

[tool.uv]
package = true

[dependency-groups]
dev = [
    "pytest-asyncio>=1.3.0",
]
